now can specify `renaming_scheme` for `ODETranslation`
* so you can control the names of the variables after translation, and so that the defaults follow those in the paper.
* allow user to specify where indexes for translated variables start.  default is at 0.
* also added a function `rename_indep_var` which does just that.
* i corrected a logic bug due to precedence rules.  and comes before or, silviana, you know this.
* corrected bug on missing substiuttion in initial conditions
* some minor code maintenance: variable names, function locations, etc.
* corrected bug on deleting subbed-away variables, but i was missing the non-constant variables, and if a dependent variable never appears on the RHS of a derivative or in any other expression, it should still be a variable and not deleted.
* improved some error messages

implementation note, also commented in the code:

I had to deal with the sympification of S --  see [this SO post](https://stackoverflow.com/questions/41860294/what-does-s-signify-in-sympy).  In `translate_parameter_substitutions`, I used `sympy.Symbol` instead of `sympy.sympify` on the new variable names , because if you give sympify just 'S', then you get `SingletonRegistry` instead of a symbol with name S.  I am not entirely satisfied with the code, but at least it gives correct results.